django rest/django_rest_api/requirements.txt
django
django-rest-framework
django_mysql
pillow
whitenoise